pyproject.toml
[project]
name = "harun_event_application"
version = "1.0.0"
description = "Harun Event Application - Event ticketing and quiz management system"
requires-python = ">=3.12"
dependencies = [
    "Flask==3.0.0",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-Mail==0.9.1",
    "Flask-WTF==1.2.1",
    "Flask-Login==0.6.3",
    "Flask-Bcrypt==1.0.1",
    "WTForms==3.1.1",
    "python-dotenv==1.0.1",
    "requests==2.31.0",
    "email-validator==2.1.0",
    "pg8000==1.31.2",
    "scramp==1.4.6",
    "redis==5.0.1",
    "qrcode==7.4.2",
    "Pillow==10.1.0",
    "reportlab==4.0.7",
    "gunicorn==21.2.0",
]

[tool.uv]
dev-dependencies = []
